[test|refactor] Extract generator for event-countable
