+++ tests/include/properties/classes/homomorphic.hpp
@@ -43,6 +43,7 @@ private:
     static constexpr HomomorphicRequirements requirements{};
 };
 
+#include "generators/values/event-countable-double-combination.hpp"
 #include "generators/values/runnable-double-combination.hpp"
 
 template <typename Expect, typename RunnableScenario, typename Then>
@@ -93,61 +94,6 @@ private:
 };
 
 template <typename EventCountable>
-struct EventCountableDoubleValueCombination
-{
-    static constexpr size_t size()
-    {
-        return std::tuple_size_v<decltype(OPS)>;
-    }
-
-    template <size_t Idx>
-    constexpr auto a() const
-    {
-        return value<A[Idx]>();
-    }
-
-    template <size_t Idx>
-    constexpr auto b() const
-    {
-        return value<B[Idx]>();
-    }
-
-private:
-    template <size_t Idx>
-    constexpr auto value() const
-    {
-        if constexpr (Idx == 0)
-        {
-            return eventCountable.prototype();
-        }
-        else
-        {
-            return OPS[Idx - 1](value<Idx - 1>());
-        }
-    }
-
-    static constexpr auto OPS = std::array<typename EventCountable::OpContainer, 8>{
-        EventCountable::template Op<COMPILE_TIME_RANDOM() % 3>(),
-        EventCountable::template Op<COMPILE_TIME_RANDOM() % 3>(),
-        EventCountable::template Op<COMPILE_TIME_RANDOM() % 3>(),
-        EventCountable::template Op<COMPILE_TIME_RANDOM() % 3>(),
-        EventCountable::template Op<COMPILE_TIME_RANDOM() % 3>(),
-        EventCountable::template Op<COMPILE_TIME_RANDOM() % 3>(),
-        EventCountable::template Op<COMPILE_TIME_RANDOM() % 3>(),
-        EventCountable::template Op<COMPILE_TIME_RANDOM() % 3>(),
-    };
-    static constexpr std::array<size_t, 8> A{
-        COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8,
-        COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8,
-    };
-    static constexpr std::array<size_t, 8> B{
-        COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8,
-        COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8, COMPILE_TIME_RANDOM() % 8,
-    };
-    static constexpr EventCountable eventCountable{};
-};
-
-template <typename EventCountable>
 struct HomomorphicEventCountableSumWithAccumulateRequirements
 {
     static constexpr size_t size()
@@ -185,7 +131,9 @@ struct HomomorphicEventCountableSumWithAccumulateRequirements
     }
 
 private:
-    static constexpr auto COMBINATIONS = EventCountableDoubleValueCombination<EventCountable>{};
+    static constexpr auto COMBINATIONS = Countable::EventCountableDoubleValueCombination<
+        EventCountable, COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM(),
+        COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM()>{};
 };
 
 template <typename EventCountable>
@@ -226,5 +174,7 @@ struct HomomorphicEventCountableSomeFailedWithAccumulateRequirements
     }
 
 private:
-    static constexpr auto COMBINATIONS = EventCountableDoubleValueCombination<EventCountable>{};
+    static constexpr auto COMBINATIONS = Countable::EventCountableDoubleValueCombination<
+        EventCountable, COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM(),
+        COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM(), COMPILE_TIME_RANDOM()>{};
 };
